fix(lx-trading-core): clippy -D warnings, and wire the arbitrage bot callback
With formatting fixed, the Rust SDK job advanced to `cargo clippy --all-targets
--all-features -- -D warnings` and failed there with 9 lib errors. Clearing those
exposed 21 more in examples/ that the lib failure had masked.

Library (9): unnecessary_sort_by x4 -> sort_by_key / Reverse; derivable_impls x2
-> #[derive(Default)] + #[default]; a hidden elided lifetime -> Timer<'_>;
field_reassign_with_default; and `VenueCapabilities::orderBook` -> `order_book`.
The `"orderBook"` at adapters/mod.rs:30 is a VENUE-NAME STRING LITERAL, not the
method, and is deliberately left alone.

Examples (21): 15 applied by rustfix (unused imports/variables, unnecessary f64
casts, clone-on-Copy). The 6 remaining were dead code, and one of them was a real
defect:

  ArbitrageBot::handle_opportunity was never called. It holds the bot's entire
  execution and stats path, so `router` was also never read and the example could
  only ever print a final report of zeros. LxFirstArbitrage::on_opportunity is the
  intended wire; start() now registers it. The callback is a sync
  Box<dyn Fn + Send + Sync>, so it spawns.

The other five were vestigial fields carrying nothing: MakerState's two
Option<Order> (the loop builds OrderRequest and simulates fills directly, so they
were only ever None) and OrderResult::symbol (every result is for the one
request's symbol). Plus a manual checked division.

Verified in the order CI runs them:

    cargo fmt --check                                CLEAN
    cargo clippy --all-targets --all-features -D     CLEAN
    cargo test --all-features                        106 passed, 0 failed
    cargo build --release                            CLEAN

+++ sdk/lx-trading-core/examples/websocket_client.rs
@@ -178,21 +178,21 @@ async fn simulate_events(tx: mpsc::Sender<WsEvent>, running: Arc<AtomicBool>) {
                     bids: vec![
                         PriceLevel::new(
                             current_price - dec!(5),
-                            Decimal::try_from(rng.gen_range(0.1..2.0) as f64).unwrap_or_default(),
+                            Decimal::try_from(rng.gen_range(0.1..2.0)).unwrap_or_default(),
                         ),
                         PriceLevel::new(
                             current_price - dec!(10),
-                            Decimal::try_from(rng.gen_range(0.5..3.0) as f64).unwrap_or_default(),
+                            Decimal::try_from(rng.gen_range(0.5..3.0)).unwrap_or_default(),
                         ),
                     ],
                     asks: vec![
                         PriceLevel::new(
                             current_price + dec!(5),
-                            Decimal::try_from(rng.gen_range(0.1..2.0) as f64).unwrap_or_default(),
+                            Decimal::try_from(rng.gen_range(0.1..2.0)).unwrap_or_default(),
                         ),
                         PriceLevel::new(
                             current_price + dec!(10),
-                            Decimal::try_from(rng.gen_range(0.5..3.0) as f64).unwrap_or_default(),
+                            Decimal::try_from(rng.gen_range(0.5..3.0)).unwrap_or_default(),
                         ),
                     ],
                     timestamp: chrono::Utc::now().timestamp_millis(),
@@ -212,12 +212,10 @@ async fn simulate_events(tx: mpsc::Sender<WsEvent>, running: Arc<AtomicBool>) {
                         Side::Sell
                     },
                     price: current_price,
-                    quantity: Decimal::try_from(rng.gen_range(0.01..0.5) as f64)
-                        .unwrap_or_default(),
+                    quantity: Decimal::try_from(rng.gen_range(0.01..0.5)).unwrap_or_default(),
                     fee: lx_trading::Fee {
                         asset: "USDC".to_string(),
-                        amount: Decimal::try_from(rng.gen_range(0.01..1.0) as f64)
-                            .unwrap_or_default(),
+                        amount: Decimal::try_from(rng.gen_range(0.01..1.0)).unwrap_or_default(),
                         rate: Some(dec!(0.001)),
                     },
                     timestamp: chrono::Utc::now().timestamp_millis(),
